Renamed auth module and updated import paths
