+++ frontend-web/src/App.tsx
@@ -1,7 +1,7 @@
 // App.tsx
 import React from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-import { useAuth } from "./firebase/useAuth";
+import { useAuth } from "./auth/useAuth";
 import Login from "./loginPage/login";
 import StudyGroupsPage from "./StudyGroupsPage/StudyGroupsPage";
 import ModuleSettings from "./moduleSettings/moduleSettings";
