+++ frontend-web/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
-import { AuthProvider } from './firebase/authProvider';
+import { AuthProvider } from './auth/authProvider';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
